Implemented Reach Routes, added Results & Details components. Clicking on logo re-directs to home
