+++ package.json
@@ -23,6 +23,7 @@
     "prettier": "^1.14.2"
   },
   "dependencies": {
+    "@reach/router": "^1.1.1",
     "petfinder-client": "0.0.3",
     "react": "^16.4.2",
     "react-dom": "^16.4.2"
