+++ src/Details.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+class Details extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>Details.js</h1>
+      </div>
+    );
+  }
+}
+
+export default Details;
