+++ src/Pet.js
@@ -1,8 +1,9 @@
 import React from "react";
+import { Link } from "@reach/router";
 
 class Pet extends React.Component {
   render() {
-    const { name, animal, breed, media, location } = this.props; //same as this.props.name & etc
+    const { name, animal, breed, media, location, id } = this.props; //same as this.props.name & etc
 
     let photos = [];
 
@@ -10,7 +11,7 @@ class Pet extends React.Component {
       photos = media.photos.photo.filter(photo => photo["@size"] === "pn");
     }
     return (
-      <div className="pet">
+      <Link to={`/details/${id}`} className="pet">
         <div className="image-container">
           <img src={photos[0].value} alt={name} />
         </div>
@@ -18,7 +19,7 @@ class Pet extends React.Component {
           <h1>{name}</h1>
           <h2>{`${animal} - ${breed} - ${location}`}</h2>
         </div>
-      </div>
+      </Link>
     );
   }
 }
